+++ src/content.ts
@@ -20,7 +20,7 @@ import {
   SignedReplyAnnouncement,
   SignedTombstoneAnnouncement,
 } from "./core/announcements";
-import { isDSNPAnnouncementURI, DSNPAnnouncementURI, InvalidAnnouncementUriError } from "./core/identifiers";
+import { isDSNPContentURI, DSNPContentURI, InvalidContentUriError } from "./core/identifiers";
 import { hash } from "./core/utilities";
 
 /**
@@ -75,19 +75,19 @@ export const broadcast = async (
  * Thrown if the from id is not configured.
  * @throws {@link InvalidActivityContentError}
  * Thrown if the provided activity content object is not valid.
- * @throws {@link InvalidAnnouncementUriError}
- * Thrown if the provided inReplyTo Announcement Uri is invalid.
+ * @throws {@link InvalidContentUriError}
+ * Thrown if the provided inReplyTo Content Uri is invalid.
  * @param contentObject - The activity content object with which to reply
- * @param inReplyTo - The DSNP Announcement Uri of the announcement that this announcement is in reply to
+ * @param inReplyTo - The DSNP Content Uri of the announcement that this announcement is in reply to
  * @param opts - Optional. Configuration overrides, such as from address, if any
  * @returns A Signed Reply Announcement ready for inclusion in a batch
  */
 export const reply = async (
   contentObject: ActivityContentNote,
-  inReplyTo: DSNPAnnouncementURI,
+  inReplyTo: DSNPContentURI,
   opts?: ConfigOpts
 ): Promise<SignedReplyAnnouncement> => {
-  if (!isDSNPAnnouncementURI(inReplyTo)) throw new InvalidAnnouncementUriError(inReplyTo);
+  if (!isDSNPContentURI(inReplyTo)) throw new InvalidContentUriError(inReplyTo);
 
   requireValidActivityContentNote(contentObject);
   const content = JSON.stringify(contentObject);
@@ -114,21 +114,21 @@ export const reply = async (
  * Thrown if the signer is not configured.
  * @throws {@link MissingFromIdConfigError}
  * Thrown if the from id is not configured.
- * @throws {@link InvalidAnnouncementUriError}
+ * @throws {@link InvalidContentUriError}
  * Thrown if the provided inReplyTo DSNP Message Id is invalid.
  * @throws {@link InvalidEmojiStringError}
  * Thrown if the emoji provided is invalid.
  * @param emoji - The emoji with which to react
- * @param inReplyTo - The DSNP Announcement Uri of the announcement to which to react
+ * @param inReplyTo - The DSNP Content Uri of the announcement to which to react
  * @param opts - Optional. Configuration overrides, such as from address, if any
  * @returns A Signed Reaction Announcement ready for inclusion in a batch
  */
 export const react = async (
   emoji: string,
-  inReplyTo: DSNPAnnouncementURI,
+  inReplyTo: DSNPContentURI,
   opts?: ConfigOpts
 ): Promise<SignedReactionAnnouncement> => {
-  if (!isDSNPAnnouncementURI(inReplyTo)) throw new InvalidAnnouncementUriError(inReplyTo);
+  if (!isDSNPContentURI(inReplyTo)) throw new InvalidContentUriError(inReplyTo);
   if (!isValidEmoji(emoji)) throw new InvalidEmojiStringError(emoji);
 
   const currentFromURI = requireGetCurrentFromURI(opts);
@@ -187,7 +187,7 @@ export const profile = async (
  * Thrown if the signer is not configured.
  * @throws {@link MissingFromIdConfigError}
  * Thrown if the from id is not configured.
- * @throws {@link InvalidAnnouncementUriError}
+ * @throws {@link InvalidContentUriError}
  * Thrown if the provided inReplyTo DSNP Message Id is invalid.
  * @throws {@link InvalidTombstoneAnnouncementTypeError}
  * Thrown if the target type provided for the tombstone is invalid.